+++ app/src/main/java/com/fazil/units/units/Units.java
@@ -4,7 +4,7 @@ public interface Units {
 
     String[] unitItems = {};
 
-    float convertQuestionToCommon(String unit, Float value);
-    float convertCommonToAnswer(String unit, Float value);
+    double convertQuestionToCommon(String unit, Double value);
+    double convertCommonToAnswer(String unit, Double value);
 
 }
